requirements.txt
biopython
tqdm
scikit-learn
pandas
transformers
torch
seaborn
